+++ test/run.sh
@@ -628,6 +628,14 @@ testGovendorGo14WithGOVERSIONOverride() {
 }
 
 testGlideMassageVendor() {
+  if [ "${IMAGE}" = "heroku/cedar:14" ]; then
+    echo "!!!"
+    echo "!!! Skipping this test on heroku/cedar:14"
+    echo "!!! It fails with gcc errors when compiling mattes/migrate"
+    echo "!!!"
+    return 0
+  fi
+
   fixture "glide-massage-vendor"
 
   env "GO_INSTALL_PACKAGE_SPEC" ". github.com/mattes/migrate"