+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add go1.11.7 and expand go1.11 to go.11.7.
 * Drop 'Go.SupportsModuleExperiment' from data.json, instead error for go versions < go1.11 when using modules.
 * Drop 'Go.Supported' from data.json since the buildpack is no longer using it for anything.
+* Skip vendored mattes migrate compile on cedar:14 due to gcc error.
 
 ## v107 (2019-04-02)
 * Handle non files in bin/ (symlinks, directories, etc) when diffing to determine contents of bin/